A graph of nodes connected by shared and weak pointers in C++
A classical example of a dynamic graph data structure, with ownership managed
by shared and weak pointers. The graph topology is like in an application that
dynamically creates sessions and registers handlers for events on each session.
Handlers keep references (smart pointers) to sessions. If all handlers for
a session are deleted (executed or canceled), the session is deleted as well.